// src/app/DiagnosticTreatment/models/UniversityDepressionIndicatorSet.ts
export class UniversityDepressionIndicatorSet{
  id: string;
  depressionPercentage: number;
  noDepressionPercentage: number;
  studentsQuantity: number;
  psychologistsQuantity: number;
  depressionLeveQuantity: number;
  depressionMinimaQuantity: number;
  depressionModeradaQuantity: number;
  depressionModeradaSeveraQuantity: number;
  depressionSeveraQuantity: number;
  universityName: any;
}